Feature/identityserver (#4)
* feat: added basic skeleton for identity service

* feat: adjusted additional config

* Revert "feat: adjusted additional config"

This reverts commit bca1e9f4d5be5da0e1a480cca9ecd0d9391824d4.

* fix: fix versioning issue for generating seed

* feat: added seed data and migration for postgres config

* feat: added skeleton for register page

* feat: added validation process for user registration

* fix: added regis html form and minor fix login layout

* feat: added client credentials for local dev

* feat: added custom profile service to extend user info in token

* feat: configured the auth endpoints on the resource server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -7,6 +7,7 @@ using AutoMapper;
 using Contracts;
 using MassTransit;
 using MassTransit.Transports;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -50,12 +51,13 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        // TODO: add current user as seller
-        auction.Seller = "test";
+
+        auction.Seller = User.Identity.Name;        // username to populate Seller property
 
         _context.Auctions.Add(auction);
 
@@ -72,6 +74,7 @@ public class AuctionsController : ControllerBase
         return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateActionDto updateActionDto)
     {
@@ -80,7 +83,8 @@ public class AuctionsController : ControllerBase
 
         if (auction == null) return NotFound();
 
-        // TODO: check seller == username
+        // http 403
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateActionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateActionDto.Model ?? auction.Item.Model;
@@ -97,6 +101,7 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes!");
     }
 
+    [Authorize]
     // For the sake of training/example
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
@@ -105,7 +110,7 @@ public class AuctionsController : ControllerBase
 
         if (auction == null) return NotFound();
 
-        // TODO: check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
